MYR-1338: fixed endpoints user-network
